requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
mysql-connector-python==8.1.0
SQLAlchemy==2.0.23
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
PyJWT==2.8.0
python-dotenv==1.0.0
pandas==2.2.0  # Newer version that might support 3.13
numpy==2.0.0
scikit-learn==1.4.0  # Try the latest version
openai==1.3.0
plotly==5.17.0
requests==2.31.0
celery==5.3.4
redis==5.0.1
gunicorn==21.2.0